Implemented timer.
I played around a bit trying to get a rect transform to shrink from the top- and I succeeded! however I realised that I couldn't figure out any method of adjusting it so that it would be complete in exactly 45 seconds (the minimum game time). Instead I went the old classic route: manipulating the slider.